✨ research pack 을 builder/renderer 책임으로 분리 추출
변경 이유:
- pack.py 가 1348 LOC 로 1000 줄 한도를 초과하고 code_audit SPLIT_NOW_PENDING
  (axes "builder, renderer") 에 등록돼 있었다. 데이터 모델 + 타입별 빌더 +
  직렬화/렌더링이 한 파일에 섞여 있어 책임 분리가 필요했다.

주요 변경 사항:
- pack_render.py 신설: pack_to_dict / pack_from_dict / pack_to_markdown 및
  _source_to_dict / _render_* 등 직렬화·마크다운 렌더 책임을 그대로 이동.
- pack_build.py 신설: source_from_* 타입별 소스 빌더 + make_research_request /
  make_finding 빌더 책임을 그대로 이동.
- pack.py 는 데이터 모델(dataclass) + URL/dedup 헬퍼 + pack 합성/머지 +
  공유 id/time 헬퍼(_now/_gen_id/_clean_url/_excerpt)만 남겨 611 LOC 로 축소.

비고:
- 순수 이동(behavior-preserving) — 로직 변경 없음.
- import 방향은 단방향: pack_build/pack_render → pack. pack 은 둘을 import 하지
  않아 순환 없음.
